+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Contributors: Gabriel Rondeau-Genesse (:user:`RondeauG`), Pascal Bourgault (:use
 Changes
 ^^^^^^^
 * Add ``max_tail_factor`` arg  to quantile mapping functions. (:pull:`279`, :issue:`278`).
+* Add ``mult_skip_zeros`` to detrending classes. (:pull:`279`, :issue:`289`).
 * New example in notebook `advanced_example.ipynb` showing how to use spectral utils (``xsdba.processing.spectral_filter`` and ``xsdba.properties.spectral_variance``). (:issue:`247`, :pull:`259`, :pull:`282`).
 
 Breaking changes